Move Navigation to new LoggedInPage Widget
